Add damage initialization endpoint and separate save path
