+++ app/api/damages/route.ts
@@ -1,4 +1,4 @@
-import { type NextRequest, NextResponse } from "next/server"
+import { NextResponse } from "next/server"
 
 const API_BASE_URL = process.env.NEXT_PUBLIC_API_URL || "https://localhost:5200/api"
 
@@ -19,30 +19,5 @@ export async function GET() {
   } catch (error) {
     console.error("Damages API error:", error)
     return NextResponse.json({ error: "Failed to fetch damages" }, { status: 500 })
-  }
-}
-
-export async function POST(request: NextRequest) {
-  try {
-    const body = await request.json()
-
-    const response = await fetch(`${API_BASE_URL}/damages`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(body),
-    })
-
-    if (!response.ok) {
-      const errorText = await response.text()
-      return NextResponse.json({ error: `Backend API error: ${errorText}` }, { status: response.status })
-    }
-
-    const data = await response.json()
-    return NextResponse.json(data)
-  } catch (error) {
-    console.error("Create damage API error:", error)
-    return NextResponse.json({ error: "Failed to create damage" }, { status: 500 })
   }
 }
